refactor: 중첩 components 폴더 정리
