+++ apps/client/app/[exhibitionId]/artwork/[artworkId]/_components/ArtistCard.tsx
