+++ apps/client/app/[exhibitionId]/artwork/_components/AlbumIcon.tsx
