+++ apps/client/app/[exhibitionId]/artwork/_components/ListIcon.tsx
